move ruff_create_syntax_error_diagnostic back to ruff
