+++ crates/ruff/src/diagnostics.rs
@@ -10,7 +10,7 @@ use std::path::Path;
 use anyhow::{Context, Result};
 use colored::Colorize;
 use log::{debug, warn};
-use ruff_db::diagnostic::ruff_create_syntax_error_diagnostic;
+use ruff_linter::message::create_syntax_error_diagnostic;
 use rustc_hash::FxHashMap;
 
 use ruff_linter::OldDiagnostic;
@@ -99,10 +99,7 @@ impl Diagnostics {
                 let name = path.map_or_else(|| "-".into(), Path::to_string_lossy);
                 let dummy = SourceFileBuilder::new(name, "").finish();
                 Self::new(
-                    vec![
-                        ruff_create_syntax_error_diagnostic(dummy, err, TextRange::default())
-                            .into(),
-                    ],
+                    vec![create_syntax_error_diagnostic(dummy, err, TextRange::default()).into()],
                     FxHashMap::default(),
                 )
             }
