+++ crates/ruff_linter/src/linter.rs
@@ -7,7 +7,7 @@ use itertools::Itertools;
 use ruff_python_parser::semantic_errors::SemanticSyntaxError;
 use rustc_hash::FxBuildHasher;
 
-use ruff_db::diagnostic::{SecondaryCode, ruff_create_syntax_error_diagnostic};
+use ruff_db::diagnostic::SecondaryCode;
 use ruff_notebook::Notebook;
 use ruff_python_ast::{ModModule, PySourceType, PythonVersion};
 use ruff_python_codegen::Stylist;
@@ -26,6 +26,7 @@ use crate::checkers::tokens::check_tokens;
 use crate::directives::Directives;
 use crate::doc_lines::{doc_lines_from_ast, doc_lines_from_tokens};
 use crate::fix::{FixResult, fix_file};
+use crate::message::create_syntax_error_diagnostic;
 use crate::noqa::add_noqa;
 use crate::package::PackageRoot;
 use crate::preview::is_py314_support_enabled;
@@ -534,20 +535,17 @@ fn diagnostics_to_messages(
     parse_errors
         .iter()
         .map(|parse_error| {
-            ruff_create_syntax_error_diagnostic(
-                source_file.clone(),
-                &parse_error.error,
-                parse_error,
-            )
-            .into()
+            create_syntax_error_diagnostic(source_file.clone(), &parse_error.error, parse_error)
+                .into()
         })
         .chain(unsupported_syntax_errors.iter().map(|syntax_error| {
-            ruff_create_syntax_error_diagnostic(source_file.clone(), syntax_error, syntax_error)
-                .into()
+            create_syntax_error_diagnostic(source_file.clone(), syntax_error, syntax_error).into()
         }))
-        .chain(semantic_syntax_errors.iter().map(|error| {
-            ruff_create_syntax_error_diagnostic(source_file.clone(), error, error).into()
-        }))
+        .chain(
+            semantic_syntax_errors.iter().map(|error| {
+                create_syntax_error_diagnostic(source_file.clone(), error, error).into()
+            }),
+        )
         .chain(diagnostics.into_iter().map(|mut diagnostic| {
             let noqa_offset = directives.noqa_line_for.resolve(diagnostic.start());
             diagnostic.set_noqa_offset(noqa_offset);
